chore: moved generate_presigned_url to response_utils.py
